Extract Repeater.js
