+++ src/CreateInterval.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = (x) => (x && (typeof x.toMs === 'function')) ? x : { toMs: function () { return x } }
